+++ feature/note/src/main/java/com/clisby/shawn/note/presentation/composables/components/notes/NoteItem.kt
@@ -1,11 +1,11 @@
-package com.clisby.shawn.note.presentation.composables.components
+package com.clisby.shawn.note.presentation.composables.components.notes
 
 import androidx.compose.foundation.layout.*
-import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
+import com.clisby.shawn.note.presentation.composables.components.NoteHeaderSubText
 import com.clisby.shawn.note.presentation.model.NoteUi
 
 @Composable
